Add a 'typefile' command
This enables a commandline like:

vncdo -s server -p password typefile abc.txt

to type the contents of abc.txt into the target server

Limitations:

This is only intended to support typeable characters, so trying to type a binary file
will fail.

CRLF pairs are emitted as a single "return" keypress, as it is expected
that the receiving OS will apply their own line endings.

An additional mapping for Tab characters was added. Other characters in
the file that are not present in the keymap are silently ignored.