Attempt to determine the correct image mode from the pixel format
VMWare's built-in VNC server seems to ignore `SetPixelFormat` messages.
However, it does correctly indicate the pixel format it uses in the
`ServerInit` message.

This commit attempts to use the pixel format info from that initial
message to derive the correct image mode to use for PIL. If that fails,
it falls back to the previous behavior of setting the pixel format to
the `rfb.py` default of `RGBX`.

This was the least-hacky way I could think of to fix VMWare screenshots,
while potentially giving some benefit to other implementations (skipping
the often unnecessary, if harmless, `SetPixelFormat` message).