Separate refreshing the screen from saving a screenshot
This change was suggested by @sibson in https://github.com/sibson/vncdotool/pull/118#issuecomment-357261582

I want to be able to use the in-memory PIL image without writing it out
to disk, through the API shim. This commit separates out the two pieces
of functionality.

What makes this a little more complicated is that you can't return a
deferred without a callback from a method on the client, or, just as if
you hadn't returned a deferred at all, the client instance gets replaced
by another object. I haven't spent the time to learn why this is the
case. To get around that, I add a no-op callback to the deferred in
`refreshScreen`.

When actually taking a screenshot, the callback to save the image is
simply added to the same deferred.