Add support for RFB versions 3.7 and 3.8
I was able to successfully connect with all three protocol versions,
with and without a password. Not providing a password when one is
required also works as expected.

However, the server I'm working with seems to closes the connection
immediately upon receiving an invalid challenge response, so I
didn't test all of the code around incorrect password handling. The
behavior is consistent with and without my patch, so isn't caused by
this change.