Make the client work with badly behaved servers
When responding to framebuffer update requests, the VMWare VNC server
will send a single-pixel update on every other response, even when a
full buffer update has been requested. To cancel this out, you can
resent any framebuffer update request that receives such a response.
Since this means essentially rejecting what are valid responses, I do
this in a subclass so as to avoid doing that by default.

The MacOS built-in VNC server is unusual in several ways.

- It advertises itself as supporting protocol version 3.889
- If no password is used it attemts a nonstandard authentication
  sequence
- 8-, 24-, and 32-bit pixel formats cause the server to disconnect as
  soon as the first post-handshake client request as made

The first two are already handled without changes to the client (it
falls back to operating in protocol version 3.8, which works fine, and
you just have to set a password), but the third requires special
handling. Since this is a nonstandard protocol version, and this only
has to happen once per session, it seems safe to build this directly
into `setImageMode`.

I currently just have my own subclass that supports these two
behaviors, but they're common enough servers (and both referenced in
open issues on this repo) that I think including the fixes here is
valuable. If you disagree, at least storing the server-sent version
number is helpful for detecting the MacOS server in a subclass.