Support an unchanged framebuffer in expectCompare
Found that expectScreen would hand if called twice in a row. This seems to be
due to the vnc server not sending an update since nothing changed.

In this particular case we should have the content already available and run one
check without waiting for the server.

If the content does not match the loop will be entered from inside
_expectCompare itself.