fix(rfb): RFB version parsing and handling
The current code just removes all zeros and handles "003.008" and
"003.080" and "003.800" all the same as `3.8`. Only because of this
broken behavior the faked header in tests/unit/test_client.py worked.

floats are also notoriously inexact, which might lead to strange errors.

Use regular expression to check for an exact match as specified in
<https://www.rfc-editor.org/rfc/rfc6143#section-7.1.1>

Use 2-tuple `[int, int]` instead of `float`.