ci: Make Python configurable
Most Linux distributions distinguish between `python2` and `python3` and
`python` defaulted to `python2` in the past. Nowadays it is frowned upon
and `python3` should be used to explicitly require Python 3.

Windows and other OS (macOS?) are latecomers and only started with
Python 3, so they only have `python` and default to Python 3.