feat: Added support for QEMU Client Messages
The RFB protocol works well for remote access to X11 applications, but
is also used by Qemu to connect to virtual machines. There the
high-level KEYSYM must be translated back to a low-level KEYCODE as VMs
get hardware events. While a keyboard layout can be be configured
with Qemu, this is imperfect as there is no unique mapping from KEYSYM
to KEYCODE.

Thus Qemu added the "Extended Key Event Message" to include they
KEYCODE, which is already provided by the X11-server running on the
client. This makes configuring they keyboard layout for Qemu obsolete.

Closes #153