feat: Add IPv6 support
While `reactor.connectTCP()` already works with IPv6 addresses, its
logic to detect IPv6 fails when an unresolved DNS name is given: In does
try to resolve the name, but has already decided to use AF_INET instead
of AF_INET6, which is only set when an IPv6 address is initially given:

> File "vncdotool/vncdotool/command.py", line 436, in vncdo():
>   factory = build_tool(options, args)
> File "vncdotool/vncdotool/command.py", line 238, in build_tool():
>   factory_connect(factory, options.host, options.port, options.address_family)
> File "vncdotool/vncdotool/client.py", line 531, in factory_connect():
>   reactor.connectTCP(host, port, factory)
> File "twisted/internet/posixbase.py", line 374, in connectTCP():
>   c.connect()
> File "twisted/internet/base.py", line 1165, in connect():
>   self.transport: Optional[Client] = self._makeTransport()
> File "twisted/internet/tcp.py", line 1517, in _makeTransport():
>   return Client(self.host, self.port, self.bindAddress, self, self.reactor)
> File "twisted/internet/tcp.py", line 714, in __init__():
>   elif abstract.isIPv6Address(host):
>     self.addressFamily = socket.AF_INET6
...
>   else:
>     self._requiresResolution = True
...
>   skt = self.createInternetSocket()
> File "twisted/internet/tcp.py", line 556, in createInternetSocket():
>   s = socket.socket(self.addressFamily, self.socketType)

Convert the code to use Twisted endpoints instead, which already has
code the iterate over all addresses returned by `getaddrinfo()` and
tries to connect to them until one works. That way a client will try
IPv6 first and fall back to IPv4:

> # getent ahosts localhost
> ::1             STREAM localhost
> 127.0.0.1       STREAM

But as noted in
<https://docs.twistedmatrix.com/en/stable/core/howto/endpoints.html>
`IStreaminClientEndpoint.connect()` expects a `Factory` only. While
`ClientFactory` is a sub-class of this, the additional methods
`clientConnectionFailed()` and `clientConnectionLost()` are not called
and requires those functions to be called explicitly from
`VNCDoToolClient.connectionLost()` and `connect().addErrback()` instead.

Switching to the newer endpoint API will hopefully also simplify adding
encryption via TLS in the future.

Closes #95